Edit view improvements (#725)
* Initial work and functionality for the View Settings dropdown in TicketsView.

* Styles for the View Settings dropdown.

* Move the list view filter call-express remove button to the left, and show it only on hover.

* Begin work on the filter-topbar redesign.

* Finish up the new design for the filter-topbar.
Add a close button to the filter-topbar.

* Style tweaks to the ticket-list table header, to match the new design.

* When duplicating a view without changing its name, add `- copy` to it.

* Use the Semantic dropdown for the view-filters operators, for a consistent look.

* Combine the basic and collection operators for tags and ticket status.

* Add a confirmation dialog when updating an existing view.

* Add the 'contains' and 'does not contain' operators to the UserSchema, to have the operators available for the Assignee filter.

* Implement dropdown selectors for Assignee, Tags and Status filters in the filter-topbar component, when editing views.
Fix broken onchange callback for the filter operator dropdowns.

* Styling for the new dropdowns in the filter-topbar.
Match the layout of the new design for the view filters.

* Don't show the edit-view-filters bar when sorting columns.
Solved by adding an extra client-only property on the view (`editMode`). Did this to keep the previous custom prop (`dirty`), needed when sending the entire view to the API for comparison.
The filter-topbar component will only render when `editMode` is true.
By default, when updating the view with the UPDATE_VIEW action, editMode turns on. You can manually turn it off. This is to keep backwards compatibility with all previous work.

* Fix lint warning with the UPDATE_VIEW action changes.

* Fix issues with wrong filter types also being rendered as dropdowns.

* Fix issues with the filter delete button showing up for both left-right filters when hovering the left filter.
The semantic ui dropdown does not get the selected value from react selects, so we need to manually update it.

* Fix issues with the `assignee` selector changing the value when getting the value from the view.
Because the current user is stored as `{current_user.id}` in the view, instead of using the id, the select wasn't able to match the active item.

* Move the list view filter call-express remove button to the left, and show it only on hover.

* Begin work on the filter-topbar redesign.

* Use the Semantic dropdown for the view-filters operators, for a consistent look.

* Implement dropdown selectors for Assignee, Tags and Status filters in the filter-topbar component, when editing views.
Fix broken onchange callback for the filter operator dropdowns.

* Styling for the new dropdowns in the filter-topbar.
Match the layout of the new design for the view filters.

* Don't show the edit-view-filters bar when sorting columns.
Solved by adding an extra client-only property on the view (`editMode`). Did this to keep the previous custom prop (`dirty`), needed when sending the entire view to the API for comparison.
The filter-topbar component will only render when `editMode` is true.
By default, when updating the view with the UPDATE_VIEW action, editMode turns on. You can manually turn it off. This is to keep backwards compatibility with all previous work.

* Fix lint warning with the UPDATE_VIEW action changes.

* Fix issues with wrong conflict resolve durring master rebase.
Fix lint warnings.