Ticket list scrollbar full height (#743)
* Make the ticket-list scrollbar take up the entire page height - instead of just the table height as previously.
Replaced the previous flexbox-based layout with absolute positioning and some javascript resizing.

* Fix lint warnings.

* Fix wrong loader positioning in the ticket-table component.
Fixed by always rendering the `.ticket-table` container, even when rendering the loader - to maintain the top spacing.

* Fix issues with the Right component in view-filters rendering both the selector and the span with the node value.
It was happening because semantic-ui's dropdown was wrapping the top-level node in RightSelect, and React wasn't able to keep track of it.

* Enter editMode when adding View filters (eg. when adding a new requester).
Render the OperatorLabel (AND/OR) in view-filters in the CallExpression component, to always stay at the right-side of the expression - like in the new filters design.

* Fix issues with broken rendering of the view filter list when there's only one call-expression left.