Refactor the complex-list table, to use divs instead of real table elements. This is usefull so we can use <a>s for table-rows, and make the ticket list items act like real links (right click actions, ctrl+click, etc.).
Fix issues with chopped checkboxes after the complex-list-table refactoring.

Support for filtering the TO column.
Filter using `name`s instead of `id`s, because ids in `messages.source.to` are different than ids from `/api/search` for the same entity.

Reduce the minimum height of the ReplyArea to 200px.
Remove the space between the editor content and the reply action, in the reply-area.

Fix issues with the reply-box jumping 2px when showing/hiding macros.

Refactoring the TicketReplyArea content layout, so that the reply area always stays the same height (depending on how much text we have in the editor), even when opening the macros.
Having some issues with draft-js not updating the text content because TicketReply is always rendered - it was previously in an IF block that was causing it to re-render when hiding the macros.

Clicking anywhere inside the reply area will focus the editor.
Required after removing the space between the editor and the applied macro action:
https://github.com/gorgias/gorgias/issues/973

Fix issues with broken macro insert and ticket message cache functionality, caused by the ticket-reply-area refactoring.
Because TicketReply is now always rendered, and not in an IF block anymore, some changes were required:
- skip the first editor change event - it triggers before we get the cached ticket message content, and overwrites it with blank content.
- use forceUpdate when inserting content from a parent/reducer (insert macro, get content from cache)

Make the TO receivers-dropdown field `required`, so you can't react a ticket without specifying at least one receiver.

Smooch integration page text tweaks: intro text, and different placeholder.

After adding a new integration, redirect to the new integration's Edit Page.
It was previously redirecting to the integrations list.

Make the source widget cards take up the full width.

Add a `fields` wrapper around the option `field`s in multi-select-field, for spacing.

Do not count deactivated integrations, in the integration type list count.

Port reply-area reafactoring to the new ticket-reply-editor component, after rebasing with master.

Render TicketReply only on `macrosVisible`, as it was previously.
Revert the changes required to have TicketReply rendered at all times.

Use react-router Link for the complex-list-table-rows.

Set complex-list-table with to auto.

Fix issues with a red-line showing up after making the receivers field `required`.

Whitelist the `<u>` tag in sanitizeHtmlDefault, to show underlined text in ticket messages.

Add `channels` to the `receiver` prop of the ticket object.
Change TO field filtering to filter by user_channel, using `receiver.channels.address`.
WIP Still has issues.

When macros are open, make the reply area 240px.

fix variable reference