Jest transition (#1814)
* Initial work on migrating to jest

* Fix failing tests after switching to jest.
Fix lint errors caused by eslint upgrade on master.

* Use enzyme instead of test-utils in the TicketMessageActions spec.
Remove expect and expect-immutable, to be able to use jest-enzyme.

* Transition all test specs that were using TestUtils to enzyme.
Remove the `react-addons-test-utils` dependency.

* Replace `expect-immutable` with `jest-immutable-matchers`.
Required for completely removing the `expect` dependency.

* Upgrade jest and jest-enzyme.
Add back the `react-addons-test-utils` dependency. Enzyme needs it util we upgrade to react@15.5.
Improve the immutable object compare in the ticket reducers spec, to work around the immutable matchers problem.

* Use Jest's default expect.
Remove the `expect` dependecy.

* Improve assertion for the click spy.

* Use Jest in test:debug.
It still has a bunch of issues:
* can't really set breakpoints
* `debugger` statements are ignored
https://github.com/facebook/jest/issues/1652

* Use Jest for test coverage.
Change the assertion in the TicketTags test, it was failing only when showing coverage. Caused by:
https://github.com/facebook/jest/issues/2007

* Remove `excluded-extensions.js`. It was only used by mocha.

* Remove jsdom. We only use enzyme shallow rendering, so it's no longer needed.