Send messages instantly (#1900)
* Initial work on sending messages instantly.
Has some issues with `receivers`. `to` receivers in the reply are broken until the api request is done.

* Improved loading style for `sending` messages.

* Remove the underscore from the `_sending` property.

* Support for errors with actions when sending messages.
Retry doesn't support actions and setting a ticket status yet.
Refactor the way we store local-only data on messages, by using an `_internal` map, to be able to store more stuff and easily clean-up before retrying.

* Support for setting the ticket status after retrying a failed ticket message submit.

* Port all existing instant-message-sendert functionality to the new state format (with newMessage), after rebasing with master.

* Fix ticketview test warning because of missing setStatus prop.

* Add missing delete-local-ticket-message reducer, for cancel-ing errored messages that are not saved in the api yet.

* Fix support for re-applying actions when retrying a failed message.

* Use `_internal.pendingMessages` for storing pending messages in the ticket state, instead of using `messages`.
Need to match the pending message for a newly posted message by comparing properties in the socket event, because we have no uid to match on. The socket responds faster than the xhr callback, so we can't use the the success reducer to remove the pending message.

* Fix issues with finding the pending messages for the newly posted ones.
Multiple pending message will now correctly resolve.

* Remove TODO about actions. Support for re-applying actions was added in e6c951b.

* Code review fixes for props usage and others.

* Fix the retry/cancel actions on errored pending messages, after moving to the `pendingMessages` state format.

* update loading animation details

* Return the rejected promise from in the submitTicketMessage action when the api call fails.
Returning the dispatch  was considered resolved.
Fixes issues with `Send & Close` messages when offline.

* Fix an issues where retried messages were placed at the end, because of updating the created_datetime on each reducer call.

* Remove subject and via from the getPendingMessageIndex message matcher.
Kept 'source' in case we two identical messages from two different users.