Keyboard navigation (#2592)
* Move the shortcutManager-specific util methods into a separate utils file.
Move the shortcutManager service into its own folder.

* Some work on focusing individual messages.

* Add a keyboard cursor for moving between a ticket messages with j/k.
Fix shortcutManager failing tests.

* Isolate the keyboard shortcuts for the messages cursor to the ticketbody component.

* Implement `f` shortcut in the ticket view, to set source type to Forward and focus editor.

* Use flow for the TicketReplyEditor component.

* Use `a` as a shortcut for opening the ticket assignee dropdown.

* Use the t key to open the tags dropdown in the tiket view.

* Close the tags and assignee dropdowns with Esc.
Improvements to the shortcutManager service, to cache each key event/action per component, and support multiple events on the same combo. Mousetrap does not support it.
Allow the 'esc' key actions to be executed all the time.
These fix the issues with the help dialog not closing with Esc, having to manually `triggerAction` to blur fields, and other issues with keyboard shortcuts.

* Use flow for the shortcutManager service.

* Improvements to the shortManager service and add some tests for it.
Update the TicketBody spec snapshot.

* Scroll to the editor when focusing it (eg. with r/f).

* Use the left/right keys to navigate between pages in ViewTable.
Create an util method to help with selecting the next page/item, and add some tests for it.
Use flow for the ViewTable/Table component.

* Implement moving between ViewTable row items with down/up or j/k.
Use flow for ViewTable/Table and ViewTable/Row.
Move the keyboard-shortcut specific to their own utils file and add tests for them.

* Select highlighted items in a ViewTable with x.

* Open selected ViewTable row item with enter.

* Bulk close tickets in ticket list with c.

* Fix issues with forcing focus on the keyboard selected item when updating a different one.
(eg. when checking a ticket, and having a different ticket selected).
Fix issues with resetting to the top item in a ViewTable list when changing page.

* Open the assignee and tags dropdowns in the ticket list with a/t.
Improvements to scrolling to the currently highlited viewtable item, to not scroll when not necessary.
Improvements to keeping the open dropdown state, to not be able to have two dropdowns open at once.
Fix broken snapshot tests for ViewTable Table and Row.

* Open the macros modal with the m key.
Autofocus the macros search field when opening the modal.
Upgrade reactstrap to 4.8 so autoFocus works correctly.

* Try to find a fix for the modal-open keyboard shortcuts unbind problem.
After the macro modal is closed, shortcuts for the components underneath are not rebound.

* Fix issues with not being able to rebind keyboard shortcuts to the component under the modal, after closing the modal.

* Update all broken jest snapshots, broken with the reactstrap upgrade.

* Use # to trash selected tickets in tickets list.

* Delete users from the user list with #.
Improvement to the popover visibility in the ticket and user list actions to work better with hotkeys.

* Implement pause functionality in the shortcutManager.
Helps with not having to manually bind/unbind components we want to ignore shortcuts from for a while (eg. when a modal is open).
Pause has a whitelist param, if we want to allow certain hotkeys/combos.

* Use alt+up/down to move between views.

* Fix issues with the navbar hotkey navigation to stop working.
Caused by sharing the namespace with the Views shortcuts, which have a different lifecycle.

* Add shift+tab to the list of shortcuts in ticket reply, for searching for macros while responding.

* Add flow to some of the files updated with the keyboard integration features.
Update broken jest snapshots.

* Split the keyboard help modal into two columns.
Move the KeyboardHelp component into it's own folder and use css modules.

* Reverse the j/k hotkeys. j is down.

* Fix performance issues with scrolling trough long lists of messages in the ticket details view.
Solved by keeping the cursor state as a private class property, and setState with a debouncer.

* Auto-focus the receivers field when setting the source type to email-forward.

* Reset the row cursor when loading new items in the viewtable table.

* Debounce the views list keyboard navigation url update.

* Reorganize and tweak the keyboard shortcuts in the help dialog.

* Rotate between the first and last items in the views list, when navigating using the alt+up/down hotkeys.

* Remove the ticket message cursor indicator.

* Tweak the text in the keyboard help dialog.

* Add a `Keyboard shortcuts` button in the dropup profile menu.

* Show the trash confirmation when deleting individual tickets.

* Add a keyboard shortcut for bulk opening tickets.

* Fix the broken navbar jest snapshot.

* Add flow to the shortcutmanager utils.

* Rename the $VARNAME nodes, to remove the $.

* Text tweak in the keyboard help dialog.

* Fix issues with the TO field stealing focus from the subject on new messages.
Use a different method for autofocusing the to field when using the `f` shortcut to set the reply to Forward.

* Add some more tests for the shortcutManager utils.

* Fix issue with missing editor toolbar icons.
Was caused by default empty array for toolbarProps.displayedActions.