Mobile improvements 2 (#2652)
* Create a separate service that handles scrolling to the active input element.
Eg. make sure the rich text area is visible when focusing the ticket reply editor.
CamelCase the TicketView classes.
Split shortcutManager utils so we can use isEditable in the new service.

* Lower the bottom padding on touch devices, to not end up on a blank screen when scrolled to the bottom.

* Fix issues with having to tap buttons with tooltips twice on touch screens (because the first tap registers as a tap).
Create a Tooltip wrapper component around reactstrap UncontrolledTooltip, to add a delay on touch screens.
Create an util file with mobile helper methods (isTouchDevice, isSmallScreen) and use it in the mobileScrollManager service.

* Rename isMobileScreen to isMobileOrSmallScreen.

* Scroll the ticket content only until we reach the bottom of the reply editor.
On touchscreens we have extra padding at the bottom, so it ended up scrolling too low.