Hide email signature (#2708)
* Initial work on separating the addSignature functionality from the setResponseText actions and reducer.

* More work on the Signature component for the reply area.
Not functional yet.

* First functional version of the add-signature button.
Still a bit buggy. Signature is not added to message if not expanded/manually-added with the button.

* Move the add signature functionality to an action and reducer.
Some code clean-up, simplify the Signature component.

* Work on automatically adding the signature when sending the message, if the signature was not manually added.

* Finish up the add signature functionality and add a minimal test and types.

* Fix issue with adding the signature at the start, when the newMessage was loaded from cache.
(eg. type message, reload page, add signature).

* Add some tests for the add signature reducer.

* More tests in actions/reducers for ticket reply, after adding the hide-signature functionality.

* Add tests for addSignature in responseUtils.

* Debug issues with newlines not being added to the signature.
It is caused by the mocked generateRandomKey, found no work-around it yet.

* Hide the signature button when the user doesn't have a signature set.
Move the hide/show signature button logic to the signature component.
Use the signature prop on the richfield only as a force toggle.

* Tests for the Signature component.

* Better defaults for checking the currentUser signature_text and _html.

* Show the signature button again after sending the message.

* Rename the Signature state var and show method names.

* Change the way we detect if the signature button should be displayed in the richfield/signature component.
Fixes issues with switching from email to internal-note and back, and with showing the add-signature button again after sending a message.
newMessage.state.dirty is true when we use setResponseText and contentState has changed (even if it doesn't contain real text or only the signature).
dirty=false after we send the message and contentState=null.

* Use the same styles for the Add Signature and Display History buttons in tickets.

* added padding-top in btn-more