Fix email signature added to chat tickets (#2745)
* Fix issues with having the same editor content when moving directly between tickets.
It was caused by the editor change debouncer. This would be trigger on unmount, and re-populate the newMessage state with the old editor value.

* Try to add some tests for the debouncer bug in the TicketReplyEditor. Not functional right now.

* Test for the TicketReplyEditor debouncer bug.
Only way to test for it is to manually trigger the change method, and mock setResponseText to be able to check the new passed value.
We can't check the new state value because the state doesn't mutate, reducers are not run in tests.
We can't check the rendered value in RichField, because it does not render even on change, and always has the initial newMessage values.

* Add flowtypes to the TicketReplyEditor component.