Ticket reply editor bugs  (#2798)
* Hide the macros if we posted another message, even if the `show_macros` preference is set.

* Add a test for the new state.firstNewMessage property.

* Hide the macros if the new message type is internal-note.
Improve the toggle macros depending on the show_macros setting method.

* Fix bug when the macros would show up again when the editor was first focused, when the "show macros by default in emails" setting was on.
If the setting was on, the macros show up by default on email messages. After closing the macros and focusing the editor (click or keyboard), the macros would show up again.
It was caused by the debounces setReponseText, that was only populating the editor content after a delay.
Fixed it by checking if the editor is focused before showing the macros.
Add a new `richFieldRef` prop on `TicketReply` that returns the `richField` ref, so we can access it from parents.
Rename `_focusEditor` to `focusEditor`, since it's a public method.

* Focus the editor after hiding the macros when selecting internal-note.

* When selecting internal-note, remove the signature from the contentState.
Add an action/reducer for removing the signature.
Still has a bug when the signature is added first, then extra text added afterwards.

* Fix issues with removing extra text added after adding the signature, when switching to internal-note.
Fixed by creating a method to search for the contentState created from the signature, within the main editor contentState.
draftjs does not have a native method for this, or a different way to search for content across multiple contentBlocks.

* Re-use the new signature content state method in the add signature method in responseUtils.
Some responseUtils clean-up.

* Tests for removing the signature from contentState.
Some clean-up in the responseUtils tests.

* Tweaks to the responseUtils and TicketReplyArea comments.
Split complicated conditional in TicketReplyArea.

* Allow manually opening macros, after they were closed on selecting internal-note.
Move the close-macros logic to the newmessage prepare action, to avoid running it on each update.

* fix macros closing when data is updated in Ticket view

(cherry picked from commit d371ac13c85024b46818c921230f344786373ba5)

* Move remove-signature functionality to the newmessage prepare action.
Removes it from the didUpdate in TicketReplyArea.
Works better and gets called only once.
Remove the removeSignature action and reducre in newMessage. We don't need them any longer.

* Use an empty contentState and selectionState in newMessage.state, instead of null.
Helps with not having to account for null state. contentState will always be a ContentState.

* Fixes for showing macros with the show_macros preference.
Improves show macros by default functionality when using the setting.

* Fix issues with adding the signature, after switching to a default empty contentState in the newMessage state.
Revert to a null selectionState, since the blank selectionState was causing problems.

* Fix issues with broken add_signature tests.
Was caused by draft-convert that removes newlines when generateRandomKey is mocked.