Go to current view keyboard shortcut (#3249)
* Add a `g v` keyboard shortcut to go the current/parent view.
If not on an item in the current active view, it stays where you are.
If there's no active view, it goes to the index page.

* Add tests for the url generator function for the active view.

* Add support for returning to a specific page and search query in the active view, when using `g v`.
Add tests for the new features.

* Fixed a bug preventing you to go back to a specific page in the search results with `g v`.
Beucase the viewtable search field triggers a change event on mount, it was removing the `page` query from the url on load.
Fix the currentPage prop in the Page component. The `fetchPage` action expects it to be a number.