Image improvements in editor (#3480)
* Create new draft-js plugin to handle pasted images in the editor.
When an image is pasted in the editor it gets inserted in the text.
The image is a dataUrl at first, but gets uploaded in the background, and the img src is replaced with the uploaded url when done.

* Share the addImage method between the draftjs plugins.

* Fix copy pasting html with images in the editor.
Images were removed previously from copy-pasted html because they were not contained in a atomic block.
Create an atomic block container in convertFromHTML when reaching an img tag.
Add the default image width in the editor in a separate config.

* Add tests for the image fixes in convertFromHTML.

* Tweak proxifyImages test to not trigger `invalid url` error.

* Make a general method for handling adding and uploading multiple image to the editor.

* Implement dragndrop images on the toolbar to add them in the text.
The ui still needs some work.

* Add files as attachments, when non-images are dragged on the toolbar.
Use flow on the toolbar component.
Rename store to toolbarStore, to not conflict with the redux store, in case we need to @connect the component.

* Improved error handling.
Show individual errors for images that didn't upload.

* Refactor handleDroppedFiles in TicketReplyEditor so we only use the dndPlugin in RicheField.

* Images dropped over the text area are inserted in text. Others are added as attachments.
Files dropped over the toolbar area, including images, are added as attachments.
Deprecate the 3rd party dnd-upload plugin, and create our own. The old was unmaintained and not configurable.
Lots of overall code improvements.

* Don't allow dragging files over the toolbar when the canDragFiles prop is not set on the editor.

* Allow dnd images to the editor, when canDropFiles is false.
Other file types are ignored.

* Rename image width to default_image_width.

* Add some comments for getEntitySelectionState.

* Remove the unused draft-js-drag-n-drop-upload-plugin package.
We now use our own custom plugin for dnd upload.

* Insert inline images only in email messages.
Improvements in the all new image editor plugins.
Use getter methods for getting parent prop values, in case we dynamically change them in the parent.

* Attach copy pasted images when adding inline images is not allowed (non-email messages).
Better error handling when inserting inline images.