Cache new message source type (#4126)
* Initial work on caching the newmessage reply source type in the ticket reply cache.

* Todos for fixing mentions support in cache.

* Cache the sourcetype in the ticket cache.
Delete the ticket cache on ticket submit.
Cache is deleted if the new message has no content.
Some issues with email-forward when got from cache.

* Save and restore mentions in the reply cache.
Because we now store the reply source type, we can store mentions as well.
Add test for addMention.
Clean-up unused remove mentions methods.

* Improve restoring the source type cache and triggering side-effects for each type.
Use the prepare action to trigger side effects for each source type when the ticket is loaded.
Eg. when we have 'email-forward' we need to set the message subject, receivers, senders and source.extra properties.