refactor(editor): extract macro search params (#989)
* refactor(editor): extract macros search query

* refactor(editor): sync search filters into hook

* refactor(editor): extract search params from `TicketReplyArea`

As a result of this I had to change a bit how the macro search works
(that will be extracted next). Before, the searchMacros function was
being called directly from various places, but now this is being done in
the `componentDidUpdate` lifecycle hook since the places that used to
call the search directly are now merely updating the query/filters.