fix(macros): include empty language if language is set (#1058)
Whenever a default language is selected in the ticket reply editor we
also want to include an empty language to mimic previous behaviour. I
had removed this because while `null` was being set previously, it was
not actually documented at all.

Now I instead went for an empty string to satisfy type constraints.