fix(ai-agent): handed over ticket event to use ai_handover tag (#6777)
* fix(ai-agent): handed over ticket event to use ai_handover tag

We can't use the assignUser action for that as it is not because AI
Agent unassigns that the ticket is Handed Over. Sometimes AI Agent
just unassigns a ticket it previously closed but that a human agent
started to interact with. We don't want to display a Handed Over
ticket event in that scenario.
Instead of using assignUser action we use the addTags (ai_handover)
action to deduce if the ticket event should be Handed Over by AI Agent
or not

* refactor: delete useless utils

* refactor: replace find by some

* refactor: encapsulate findAndSplitMessageTags