refactor(translations): pass translations as nested objected instead of merging them into the message object (#10367)
* refactor(translations): pass translations as nested objected instead of
merging them into the message object

This change enable us to preserve the validity of the isStripped check
in the Content component which is needed to know if we need to show the
Ellipsis allowing the user to show the expanded version of the message.

Previously, merging the translations data into the Message object would
make the strippedContent value always different from the content value
in the Content component.\
\
This commits fix that issue.

* feat: handle edge case where the message doesn't have a stripped_html or stripped_text values but has a body_html & body_text values (like in facebook messages)

* feat: rm comments